exfat support
exFAT is a proposed replacement for FAT32 that addresses many of its limitations. One of the major flaws of FAT32 is the rapid fragmentation of files. This is because the constant creation and deletion of files creates noncontiguous areas of free space. When large files are placed into these spaces, they are broken into many pieces to fit. exFAT uses a free space bitmap just like other modern file systems. This allows exFAT to find contiguous space where the file will fit, thereby reducing the occurrence of fragmentation. @DooMLoRD @faux123

Signed-off-by: Rachit Rawat <rachit.rc96@gmail.com>